r/storage_account_network_rules: refactoring to use `hashicorp/go-azure-sdk`
